requirements.txt
PyYAML==6.0.1
Flask==2.2.2
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
pandas==2.1.3
scikit-learn==1.3.2
xgboost==2.0.2
matplotlib==3.8.2
xarray==2023.12.0
gunicorn==21.2.0

# don't change the order start from here
ecmwflibs==0.5.3
eccodes==1.6.1
cfgrib==0.9.10.4


